updated for python 3 in R23
The plugin started throwing errors due to the omission of brackets everywhere there was "print". Fixed this, also changed an iterator to be Python 2/3 compatible, as opposed to only Python 2. There are a few new FBX export preferences, which means the "override FBX export for this object" option likely needs updatding. I'm thinking about ditching this option altogether, as I never use it. I'm not really sure why I put it in there in the first place.